Refractor: course model to course app
