+++ elearning/apps/accountapp/models.py
@@ -15,13 +15,6 @@ class Instructor(models.Model):
     profile_img = models.ImageField(null=True, blank=True, upload_to="images/") #default image location
     
     
-class Course(models.Model):
-    instructor = models.ForeignKey(Instructor,on_delete=models.CASCADE)
-    title = models.CharField(max_length=100)
-    description = models.TextField()
-    certificate = models.ImageField(null=True,blank=True,upload_to="images/")
-    date_created = models.DateTimeField(auto_now=True) #get current date time once it created
-    
 class Student(models.Model):
     #get random characters
     def RandomStudentNo():
